+++ tests/commands/test_organizations.py
@@ -464,7 +464,7 @@ def test_organizations_secrets_list(mock_request):
 
 
 @mock.patch.object(Client, "request", return_value=({}, None))
-def test_organizations_secrets_create(mock_request):
+def test_organizations_secrets_aws_create(mock_request):
     org_id = gen_uuid()
     name = "my_secret"
     secret_type = "AWS"
@@ -495,6 +495,39 @@ def test_organizations_secrets_create(mock_request):
             "name": name,
             "type": secret_type,
             "data": {"access_key": access_key, "secret_key": secret_key},
+        },
+    )
+
+
+@mock.patch.object(Client, "request", return_value=({}, None))
+def test_organizations_secrets_azure_create(mock_request):
+    org_id = gen_uuid()
+    name = "my_secret"
+    secret_type = "azure"
+    connection_string = "my_connection_string"
+
+    call_command(
+        "croud",
+        "organizations",
+        "secrets",
+        "create",
+        "--org-id",
+        org_id,
+        "--name",
+        name,
+        "--type",
+        secret_type,
+        "--connection-string",
+        connection_string,
+    )
+    assert_rest(
+        mock_request,
+        RequestMethod.POST,
+        f"/api/v2/organizations/{org_id}/secrets/",
+        body={
+            "name": name,
+            "type": secret_type,
+            "data": {"connection_string": connection_string},
         },
     )